refactor: feature_login pull함
