+++ thisplay/src/main/java/com/example/thisplay/global/util/PasswordValidator.java
@@ -1,4 +1,4 @@
-package com.example.thisplay.util;
+package com.example.thisplay.global.util;
 
 import java.util.regex.Pattern;
